moved folders and installed jest
