+++ package.json
@@ -6,8 +6,8 @@
   "scripts": {
     "start": "node index.js",
     "watch": "nodemon index.js",
-    "dev": "ts-node ./src-new/server.ts",
-    "lint": "eslint ./src-new --ext .ts,.tsx --fix"
+    "dev": "ts-node ./src/server.ts",
+    "lint": "eslint ./src --ext .ts,.tsx --fix"
   },
   "author": "Shayne Stewart",
   "license": "ISC",
@@ -23,15 +23,18 @@
   "devDependencies": {
     "@types/body-parser": "^1.17.0",
     "@types/express": "^4.16.1",
+    "@types/jest": "^24.0.12",
     "@typescript-eslint/eslint-plugin": "^1.7.0",
     "@typescript-eslint/parser": "^1.7.0",
     "eslint": "^5.16.0",
     "eslint-config-prettier": "^4.2.0",
     "eslint-plugin-prettier": "^3.0.1",
     "husky": "^2.2.0",
+    "jest": "^24.8.0",
     "lint-staged": "^8.1.6",
     "nodemon": "^1.19.0",
-    "prettier": "^1.17.0"
+    "prettier": "^1.17.0",
+    "ts-jest": "^24.0.2"
   },
   "husky": {
     "hooks": {
