--- src-new/server.ts
@@ -1,5 +1,0 @@
-import App from './app';
-
-const app = new App(3000);
-
-app.listen();
