--- src/authentication/logIn.dto.ts
@@ -1,11 +1,0 @@
-import { IsString } from "class-validator";
-
-class LogInDto {
-  @IsString()
-  public email: string;
-
-  @IsString()
-  public password: string;
-}
-
-export default LogInDto;
