--- src/exceptions/AuthenticationTokenMissingException.ts
@@ -1,9 +1,0 @@
-import HttpException from "./HttpException";
-
-class AuthenticationTokenMissingException extends HttpException {
-  constructor() {
-    super(401, "Authentication token missing");
-  }
-}
-
-export default AuthenticationTokenMissingException;
