--- src/exceptions/HttpException.ts
@@ -1,11 +1,0 @@
-class HttpException extends Error {
-  status: number;
-  message: string;
-  constructor(status: number, message: string) {
-    super(message);
-    this.status = status;
-    this.message = message;
-  }
-}
-
-export default HttpException;
