--- src/exceptions/PostNotFoundException.ts
@@ -1,9 +1,0 @@
-import HttpException from "./HttpException";
-
-class PostNotFoundException extends HttpException {
-  constructor(id: string) {
-    super(404, `Post with id ${id} not found`);
-  }
-}
-
-export default PostNotFoundException;
