--- src/exceptions/UserWithThatEmailAlreadyExistsException.ts
@@ -1,9 +1,0 @@
-import HttpException from "./HttpException";
-
-class UserWithThatEmailAlreadyExistsException extends HttpException {
-  constructor(email: string) {
-    super(400, `User with email ${email} already exists`);
-  }
-}
-
-export default UserWithThatEmailAlreadyExistsException;
