--- src/exceptions/WrongCredentialsException.ts
@@ -1,9 +1,0 @@
-import HttpException from "./HttpException";
-
-class WrongCredentialsException extends HttpException {
-  constructor() {
-    super(401, "Wrong credentials provided");
-  }
-}
-
-export default WrongCredentialsException;
