--- src/interfaces/controller.interface.ts
@@ -1,8 +1,0 @@
-import { Router } from "express";
-
-interface Controller {
-  path: string;
-  router: Router;
-}
-
-export default Controller;
