--- src/middleware/error.middleware.ts
@@ -1,18 +1,0 @@
-import { NextFunction, Request, Response } from "express";
-import HttpException from "../exceptions/HttpException";
-
-function errorMiddleware(
-  error: HttpException,
-  request: Request,
-  response: Response,
-  next: NextFunction
-) {
-  const status = error.status || 500;
-  const message = error.message || "Something went wrong";
-  response.status(status).send({
-    status,
-    message,
-  });
-}
-
-export default errorMiddleware;
