--- src/posts/post.dto.ts
@@ -1,11 +1,0 @@
-import { IsString } from "class-validator";
-
-class CreatePostDto {
-  @IsString()
-  public content: string;
-
-  @IsString()
-  public title: string;
-}
-
-export default CreatePostDto;
