--- src/posts/post.interface.ts
@@ -1,7 +1,0 @@
-interface Post {
-  authorId: string;
-  content: string;
-  title: string;
-}
-
-export default Post;
