--- src/users/user.dto.ts
@@ -1,14 +1,0 @@
-import { IsString } from "class-validator";
-
-class CreateUserDto {
-  @IsString()
-  public name: string;
-
-  @IsString()
-  public email: string;
-
-  @IsString()
-  public password: string;
-}
-
-export default CreateUserDto;
