--- src/users/user.interface.ts
@@ -1,8 +1,0 @@
-interface User {
-  _id: string;
-  name: string;
-  email: string;
-  password: string;
-}
-
-export default User;
