--- src/utils/validateEnv.ts
@@ -1,13 +1,0 @@
-import { cleanEnv, port, str } from "envalid";
-
-function validateEnv() {
-  cleanEnv(process.env, {
-    JWT_SECRET: str(),
-    MONGO_PASSWORD: str(),
-    MONGO_PATH: str(),
-    MONGO_USER: str(),
-    PORT: port(),
-  });
-}
-
-export default validateEnv;
